- исправление замечаний

+++ gateway/src/main/java/ru/practicum/shareit/booking/dto/BookItemRequestDto.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class BookItemRequestDto {
 
-	private long itemId;
+	private int itemId;
 
 	@FutureOrPresent
 	private LocalDateTime start;
